- applied Matthias patch
please find attached a patch against WSim revision 302. It fixes some stability
(security) issues and types. The description is inside the patch file.

This patch fixes important security issues around the ctype.h is*() and
to*() functions.  Note that these functions (1) expect an int argument
(because they need to map all unsigned chars, plus EOF == -1) and
(2) are implemented as array on many operating systems, where characters
in the range 0x80..0xff can wreak havoc, because they will be sign
extended to 0xf...f80 .. 0xf...fff, or -128 to -1 (where the latter
aliases EOF).

Hence, to disambiguate the arguments for accented characters, we need to
case the to*() and is*() argument to unsigned char unconditionally so
that it remains positive 128...255 after sign extension.

Luckily, this triggers "warning: subscript has type `char'" complaints
at least on Cygwin and Solaris, so it was easy to spot.


Also fix formatting (PRIxPTR is %lx on Cygwin 1.7 and hence doesn't
match uintptr_t which is unsigned long -- use unsigned long and %lx
explicitly; fixing related compiler warnings) and a typo (verbode ->
verbose).


git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/wsim@304 c0ef0dea-404d-0410-ad65-84cceb5f200a